requirements.txt
scikit-learn
pillow
joblib
numpy
scikit-image
opencv-python-headless
opencv-contrib-python